requirements.txt
pytaglib~=1.4.6
spotipy~=2.16.1